V Done
Added code to football.c and temperature.c using an online C compiler because I was so tired I forgot I was supposed to be programming using linux and couldn't get my desktop to compile C. Whoops. That's what I get for spending too much time in the hospital and at the gym.